// apartment-frontend/src/app/model/apartment.ts
export class Apartment {
  idApartment: number;
  title: string;
  nameApartment: string;

  room: number;
  baths: number;

  area: number;
  price: number;
  description: string;
  listingStatus: number;
  rentalStatus: number;
  streetName: string;
  idWard: number;

  postingDate: Date;
  endDate: Date;
  idAccount: number;
  idType: number;
  heartCount: number;
  img: string;
}